rename test to testing
