+++ testing/mock-server/stubs/CreateNamespace.json
