+++ testing/mock-server/stubs/DeleteNamespace.json
