+++ testing/mock-server/stubs/GetNVMfRemoteController.json
